Bug: update pseudoDEG() and createPB() to reflect Seurat v5 requirments
After considering all options this change retained the lfcThreshold specification
when calling the results() function in the DESeq workflow. This markedly reduces
the number of DEGs identified relative to the default value of 0. I may opt in
the future to use a value of 0, then post hoc filter the data.

In addition to adjusting this work flow, and option to run the LRT test instead
of Wald was added and can be specifed with the test.use formal.