Update integrateData and plotgsea
There was an error in integrateData, such that a sample would only be removed if
less than 30 cells. This was suppoosed to be a used defined value.

Add lollipop plot option for plotgsea function.

Update crossSpeciesDEG functionto crossConditionDEG for broader applicability.

Add sigDEG_heatmap for data vis of identified DEGs.